+++ core/data/src/main/java/com/afternote/core/data/di/CoreRepositoryModule.kt
@@ -1,6 +1,8 @@
 package com.afternote.core.data.di
 
+import com.afternote.core.data.repositoryImpl.auth.AccountRepositoryImpl
 import com.afternote.core.data.repositoryImpl.auth.AuthRepositoryImpl
+import com.afternote.core.domain.repository.AccountRepository
 import com.afternote.core.domain.repository.AuthRepository
 import dagger.Binds
 import dagger.Module
@@ -10,8 +12,12 @@ import javax.inject.Singleton
 
 @InstallIn(SingletonComponent::class)
 @Module
-fun interface CoreRepositoryModule {
+interface CoreRepositoryModule {
     @Binds
     @Singleton
     fun bindAuthRepository(impl: AuthRepositoryImpl): AuthRepository
+
+    @Binds
+    @Singleton
+    fun bindAccountRepository(impl: AccountRepositoryImpl): AccountRepository
 }
