+++ core/domain/src/main/java/com/afternote/core/domain/repository/AuthRepository.kt
@@ -1,9 +1,7 @@
 package com.afternote.core.domain.repository
 
-import com.afternote.core.model.EmailVerifyResult
 import com.afternote.core.model.LoginResult
 import com.afternote.core.model.ReissueResult
-import com.afternote.core.model.SignUpResult
 import kotlinx.coroutines.flow.Flow
 
 interface AuthRepository {
@@ -29,20 +27,6 @@ interface AuthRepository {
     suspend fun getUserId(): Long?
     // 레거시 레포 기준
 
-    suspend fun sendEmailCode(email: String): Result<Unit>
-
-    suspend fun verifyEmail(
-        email: String,
-        certificateCode: String,
-    ): Result<EmailVerifyResult>
-
-    suspend fun signUp(
-        email: String,
-        password: String,
-        name: String,
-        profileUrl: String?,
-    ): Result<SignUpResult>
-
     suspend fun login(
         email: String,
         password: String,
@@ -53,9 +37,4 @@ interface AuthRepository {
     suspend fun reissue(refreshToken: String): Result<ReissueResult>
 
     suspend fun logout(refreshToken: String): Result<Unit>
-
-    suspend fun passwordChange(
-        currentPassword: String,
-        newPassword: String,
-    ): Result<Unit>
 }
