+++ feature/afternote/data/src/main/kotlin/com/afternote/feature/afternote/data/repositoryimpl/MemorialThumbnailUploadRepositoryImpl.kt
@@ -1,8 +1,8 @@
 package com.afternote.feature.afternote.data.repositoryimpl
 
-import com.afternote.core.network.ImageApiService
 import com.afternote.core.network.dto.PresignedUrlRequestDto
 import com.afternote.core.network.model.requireData
+import com.afternote.core.network.service.ImageApiService
 import com.afternote.feature.afternote.domain.repository.MemorialThumbnailUploadRepository
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.withContext
